+++ Src/GBX.NET/GameBoxHeaderOfT.cs
